Extract questioned/reference/background samples from a list of populations from sources

The function splits a list of items (rows) into a sample of reference items, questioned items and background items.

make_idx_splits(sources, k_ref, k_quest, source_ref = NULL,
  source_quest = NULL, same_source = NULL, background = "outside",
  replace = TRUE)

Arguments

sources
all class labels

k_ref
number of reference samples

k_quest
number of questioned samples

source_ref
the reference source (scalar; if NULL, a random source will be picked)

source_quest
the questioned source(s) (if NULL, anything but the reference source: behaviour overridden by same_source)

same_source
if source_quest is NULL and same_source is TRUE, questioned source is the reference source

background
see details (default: 'outside')

replace
use sampling with replacement, else error

Value

list of indexes (idx_reference, idx_questioned, idx_background)

Details

Reference/questioned/background samples are always non-intersecting, even when the source is the same.

Sampling with replacement is used, if necessary and not forbidden. If it is used, a message appears.

It is always guaranteed that no sample appear more than once across reference/questioned/background items (but it can appear multiple times in a set).

Source sampling

If source_quest is NULL:

if same_source is NULL or FALSE, questioned items are sampled from all but the reference source.

if same_source is TRUE, questioned items are sampled from the reference source.

Else, questioned items will be sampled from the questioned source(s), even if it contains the reference one.

Items will never be sampled once (unless replace is TRUE): they appear once in the reference/questioned/background items.

Background selection

If background is 'outside', the background dataset comprises all items who do not lie in any of the reference and questioned sets. It can contain items from reference and questioned sources.

If background is 'others', the background dataset comprises all items from the non-reference and non-questioned sources.

By default, background is 'outside'.

Notice that background = 'others' generates no background data if questioned sources are not specified: the union of reference and questioned sources fills the available sources in the population.

See also